requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
transformers==4.35.2
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.4.2
numpy==1.26.2